+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced hard-coded color values.
 - Show quoted notes in note cards.
 - Removed wss:// from relay addresses in lists and removed the need to prepend relay addresses with wss://.
+- Localized the quotation marks on the Notifications view.
 
 ## [0.1.25] - 2024-08-21Z